move modal sources: vm-migrate
